+++ Assets/Scripts/Source/Game/Enemy.cs
@@ -1,6 +1,6 @@
 public class Enemy : Character 
 {
-    public Enemy(string name, int health, float speed) : base(name, health, speed)
+    public Enemy(string name, int health, int speed) : base(name, health, speed)
     {
 
     }
